Initial jobqueue plugin which adds jobqueue_callback method to app
